MFC r367530: RFC 7323 specifies that: * TCP segments without timestamps should be dropped when support for   the timestamp option has been negotiated. * TCP segments with timestamps should be processed normally if support   for the timestamp option has not been negotiated. This patch enforces the above. Manually resolved merge conflicts.
MFC 367891:
Fix an issue I introuced in r367530: tcp_twcheck() can be called
with to == NULL for SYN segments. So don't assume tp != NULL.
Thanks to jhb@ for reporting and suggesting a fix.

MFC r367946:
Fix two occurences of a typo in a comment introduced in r367530.
Thanks to lstewart@ for reporting them.

PR:			250499
Reviewed by:		gnn, rrs
Sponsored by:		Netflix, Inc.
Differential Revision:	https://reviews.freebsd.org/D27148


git-svn-id: https://svn.freebsd.org/base/stable/12@368181 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f